+++ dvha/db/sql_connector.py
@@ -19,7 +19,7 @@ from os.path import dirname, join, isfile
 from dvha.db.sql_columns import categorical, numerical
 from dvha.options import Options
 from dvha.paths import CREATE_PGSQL_TABLES, CREATE_SQLITE_TABLES, DATA_DIR
-from dvha.tools.errors import SQLError
+from dvha.tools.errors import SQLError, push_to_log
 import json
 
 
@@ -193,7 +193,7 @@ class DVH_SQL:
         :param column: SQL column to be updated
         :type column: str
         :param value: value to be set
-        :type value: str
+        :type value: str or float or int
         :param condition_str: a condition in SQL syntax
         :type condition_str: str
         """
@@ -429,7 +429,7 @@ class DVH_SQL:
 
     def is_sql_table_empty(self, table):
         """
-        Check if specifed SQL table is empty
+        Check if specified SQL table is empty
         :param table: SQL table
         :type table: str
         :return: True if specified table has no data
@@ -713,7 +713,7 @@ def echo_sql_db(config=None, db_type='pgsql', group=1):
         return True
     except Exception as e:
         if type(e) not in [psycopg2.OperationalError, sqlite3.OperationalError]:
-            print(str(e))
+            push_to_log(e, msg='Unknown Error during SQL Echo')
         return False
 
 
@@ -726,10 +726,14 @@ def write_test(config=None, db_type='pgsql', group=1, table=None, column=None, v
             cnx = DVH_SQL(config, db_type=db_type, group=group)
         else:
             cnx = DVH_SQL(group=group)
+    except Exception as e:
+        push_to_log(e, msg="Write Test: Connection to SQL could not be established")
+        return {'write': False, 'delete': False}
+
+    try:
         cnx.initialize_database()
     except Exception as e:
-        print("%s" % e)
-        return {'write': False, 'delete': False}
+        push_to_log(e, msg="Write Test: DVH_SQL.initialize_database failed")
 
     if table is None:
         table = cnx.tables[-1]
@@ -751,18 +755,33 @@ def write_test(config=None, db_type='pgsql', group=1, table=None, column=None, v
     insert_cmd = "INSERT INTO %s (%s) VALUES ('%s');" % (table, column, value)
     delete_cmd = "DELETE FROM %s WHERE %s;" % (table, condition_str)
 
-    cnx.execute_str(insert_cmd)
-    test_return = cnx.query(table, column, condition_str)
-    write_test_success = len(test_return) > 0
+    try:
+        cnx.execute_str(insert_cmd)
+    except Exception as e:
+        push_to_log(e, msg="Write Test: SQL test insert command failed")
+
+    try:
+        test_return = cnx.query(table, column, condition_str)
+        write_test_success = len(test_return) > 0
+    except Exception as e:
+        write_test_success = False
+        push_to_log(e, msg="Write Test: SQL query of test insert failed")
 
     if not write_test_success:
         delete_test_success = None
     else:
-        cnx.execute_str(delete_cmd)
-        test_return = cnx.query(table, column, condition_str)
-        delete_test_success = len(test_return) == 0
+        try:
+            cnx.execute_str(delete_cmd)
+            test_return = cnx.query(table, column, condition_str)
+            delete_test_success = len(test_return) == 0
+        except Exception as e:
+            delete_test_success = False
+            push_to_log(e, msg="Write Test: SQL delete command of test insert failed")
 
-    cnx.close()
+    try:
+        cnx.close()
+    except Exception as e:
+        push_to_log(e, msg='Write Test: Failed to close SQL connection')
 
     return {'write': write_test_success, 'delete': delete_test_success}