+++ dvha/models/plot.py
@@ -25,7 +25,7 @@ from os.path import join, isdir, splitext
 from os import mkdir
 from scipy.stats import ttest_ind, ranksums, normaltest
 from dvha.dialogs.export import save_data_to_file
-from dvha.tools.errors import PlottingMemoryError, ErrorDialog
+from dvha.tools.errors import PlottingMemoryError, ErrorDialog, push_to_log
 from dvha.tools.utilities import collapse_into_single_dates, moving_avg, is_windows, FIG_WILDCARDS
 from dvha.tools.stats import MultiVariableRegression, get_control_limits
 from dvha.paths import TEMP_DIR
@@ -130,9 +130,9 @@ class Plot:
     def update_bokeh_layout_in_wx_python(self):
         try:
             self.html_str = get_layout_html(self.bokeh_layout)
-        except MemoryError:
-            print('ERROR: dvha.models.plot in Plot.update_bokeh_layout_in_wx_python with '
-                  'bokeh.io.export.get_layout_html() raised MemoryError')
+        except MemoryError as e:
+            msg = 'Plot.update_bokeh_layout_in_wx_python: bokeh.io.export.get_layout_html failed'
+            push_to_log(e, msg=msg)
             raise PlottingMemoryError(self.type)
         if is_windows():  # Windows requires LoadURL()
             if not isdir(TEMP_DIR):
@@ -270,7 +270,8 @@ class Plot:
         try:
             file_name = save_data_to_file(parent, title, partial(self.save_figure, attr_dicts),
                                           initial_dir="", data_type='function', wildcard=FIG_WILDCARDS)
-            ErrorDialog(parent, "Output saved to %s" % file_name, "Save Successful", flags=wx.OK | wx.OK_DEFAULT)
+            if file_name is not None:
+                ErrorDialog(parent, "Output saved to %s" % file_name, "Save Successful", flags=wx.OK | wx.OK_DEFAULT)
         except Exception as e:
             if "phantomjs is not present" in str(e).lower():
                 msg = "Please download a phantomjs executable from https://phantomjs.org/download.html " \
@@ -724,7 +725,8 @@ class PlotTimeSeries(Plot):
                     s[grp], p[grp] = normaltest(data)
                     p[grp] = "%0.3f" % p[grp]
                 except Exception as e:
-                    print('Normal test failed: ', str(e))
+                    msg = 'PlotTimeSeries.update_divs: Normal test failed'
+                    push_to_log(e, msg=msg)
                     p[grp] = 'ERROR'
 
         # t-Test and Rank Sums
@@ -734,13 +736,15 @@ class PlotTimeSeries(Plot):
                 st, pt = ttest_ind(grp_data[1], grp_data[2])
                 pt = "%0.3f" % pt
             except Exception as e:
-                print('t-Test failed: ', str(e))
+                msg = 'PlotTimeSeries.update_divs: t-test failed'
+                push_to_log(e, msg=msg)
                 pt = 'ERROR'
             try:
                 sr, pr = ranksums(grp_data[1], grp_data[2])
                 pr = "%0.3f" % pr
             except Exception as e:
-                print('Wilcoxon ranksums failed: ', str(e))
+                msg = 'PlotTimeSeries.update_divs: Wilcoxon ranksums failed'
+                push_to_log(e, msg=msg)
                 pr = 'ERROR'
 
         self.normal_test_div[1].text = "<b>Group 1 Normal Test p-value</b>: %s" % p[1]
@@ -1591,7 +1595,7 @@ class PlotControlChart(Plot):
         self.adj_figure.plot_width = int(self.size_factor['plot'][0] * float(panel_width))
         self.adj_figure.plot_height = int(self.size_factor['plot'][1] * float(panel_height))
 
-    def update_plot(self, x, y, mrn, uid, dates, y_axis_label='Y Axis', update_layout=True):
+    def update_plot(self, x, y, mrn, uid, dates, y_axis_label='Y Axis', update_layout=True, cl_overrides=None):
         self.set_figure_dimensions()
         self.clear_sources()
         self.y_axis_label = y_axis_label
@@ -1600,14 +1604,19 @@ class PlotControlChart(Plot):
         x, y, mrn, uid, dates = self.clean_data(x, y, mrn=mrn, uid=uid, dates=dates)
 
         center_line, ucl, lcl = get_control_limits(y)
+        if cl_overrides is not None:
+            if 'UCL' in cl_overrides and cl_overrides['UCL'] is not None:
+                ucl = cl_overrides['UCL']
+            if 'LCL' in cl_overrides and cl_overrides['LCL'] is not None:
+                lcl = cl_overrides['LCL']
 
         plot_color = [self.options.PLOT_COLOR_2, self.options.PLOT_COLOR][self.group == 1]
         ooc_color = [self.options.CONTROL_CHART_OUT_OF_CONTROL_COLOR_2,
                      self.options.CONTROL_CHART_OUT_OF_CONTROL_COLOR][self.group == 1]
         colors = [ooc_color, plot_color]
         alphas = [self.options.CONTROL_CHART_OUT_OF_CONTROL_ALPHA, self.options.CONTROL_CHART_CIRCLE_ALPHA]
-        color = [colors[ucl > value > lcl] for value in y]
-        alpha = [alphas[ucl > value > lcl] for value in y]
+        color = [colors[ucl >= value >= lcl] for value in y]
+        alpha = [alphas[ucl >= value >= lcl] for value in y]
 
         self.source['plot'].data = {'x': x, 'y': y, 'mrn': mrn, 'uid': uid,
                                     'color': color, 'alpha': alpha, 'dates': dates}