+++ dvha/options.py
@@ -17,6 +17,7 @@ import hashlib
 from copy import deepcopy
 from dvha._version import __version__
 from dvha.paths import OPTIONS_PATH, OPTIONS_CHECKSUM_PATH, INBOX_DIR, IMPORTED_DIR, REVIEW_DIR
+from dvha.tools.errors import push_to_log
 
 
 class DefaultOptions:
@@ -41,9 +42,6 @@ class DefaultOptions:
                           'DB_TYPE_GRPS', 'SQL_LAST_CNX_GRPS']
 
         self.MIN_BORDER = 50
-
-        self.MAX_FIELD_SIZE_X = 400  # in mm
-        self.MAX_FIELD_SIZE_Y = 400  # in mm
 
         # These colors propagate to all tabs that visualize your two groups
         self.PLOT_COLOR = 'blue'
@@ -173,8 +171,10 @@ class DefaultOptions:
         # This is the number of bins up do 100% used when resampling a DVH to fractional dose
         self.RESAMPLED_DVH_BIN_COUNT = 5000
 
-        self.COMPLEXITY_SCORE_X_WEIGHT = 1.
-        self.COMPLEXITY_SCORE_Y_WEIGHT = 1.
+        self.MLC_ANALYZER_OPTIONS = {'max_field_size_x': 400.,
+                                     'max_field_size_y': 400.,
+                                     'complexity_weight_x': 1.,
+                                     'complexity_weight_y': 1.}
 
         # Per TG-263 (plus NONE, ITV, and IGNORED)
         self.ROI_TYPES = ['NONE', 'ORGAN', 'PTV', 'ITV', 'CTV', 'GTV',
@@ -237,8 +237,8 @@ class Options(DefaultOptions):
                     loaded_options = pickle.load(infile)
                 self.upgrade_options(loaded_options)
             except Exception as e:
-                print('ERROR: Options file corrupted. Loading default options.')
-                print("%s" % e)
+                msg = 'Options.load: Options file corrupted. Loading default options.'
+                push_to_log(e, msg=msg)
                 loaded_options = {}
 
             for key, value in loaded_options.items():
@@ -263,7 +263,8 @@ class Options(DefaultOptions):
         :param value: value of option
         """
         if not hasattr(self, attr):
-            print('WARNING: This option did not previously exist!')
+            msg = 'Options.set_option: %s did not previously exist' % attr
+            push_to_log(msg=msg)
         setattr(self, attr, value)
         self.is_edited = True
 
@@ -296,8 +297,9 @@ class Options(DefaultOptions):
             stored_checksum = self.load_stored_checksum()
             if current_checksum == stored_checksum:
                 return True
-        except Exception:
-            print('Corrupted options file detected. Loading default options.')
+        except Exception as e:
+            msg = 'Options.is_options_file_valid: Corrupted options file detected. Loading default options.'
+            push_to_log(e, msg=msg)
             return False
 
     def restore_defaults(self):